# perf: make some tiny optimizations to ResumePlayerSessionAction