--- a/app/Platform/Actions/ResumePlayerSessionAction.php
+++ b/app/Platform/Actions/ResumePlayerSessionAction.php
@@ -58,11 +58,15 @@ class ResumePlayerSessionAction
             ->orderByDesc('id')
             ->first();
 
+        // As best as we can, we'll try to only write once to the user model.
+        $doesUserNeedsUpdate = false;
+
         if ($user->LastGameID !== $game->id) {
             expireRecentlyPlayedGames($user->User);
+
             // TODO deprecated, read from last player_sessions entry where needed
             $user->LastGameID = $game->id;
-            $user->saveQuietly();
+            $doesUserNeedsUpdate = true;
         }
 
         // if the session is less than 10 minutes old, resume session
@@ -91,7 +95,7 @@ class ResumePlayerSessionAction
                 // TODO deprecated, read from last player_sessions entry where needed
                 $user->RichPresenceMsg = utf8_sanitize($presence);
                 $user->RichPresenceMsgDate = $timestamp;
-                $user->saveQuietly();
+                $doesUserNeedsUpdate = true;
 
                 // Update the player's game_recent_players table entry.
                 GameRecentPlayer::upsert(
@@ -120,6 +124,10 @@ class ResumePlayerSessionAction
             }
 
             $playerSession->save(['touch' => true]);
+
+            if ($doesUserNeedsUpdate) {
+                $user->saveQuietly();
+            }
 
             PlayerSessionResumed::dispatch($user, $game, $presence);