This PR implements small optimizations to `ResumePlayerSessionAction`. Individually, these optimizations only save 3-5ms per execution, but at 49M executions/month they should yield significant CPU time savings.

**Changes:**
- `User` model updates have been consolidated from 3 separate saves to a single save.
- `AttachPlayerGameAction` now uses `firstOrCreate()` instead of a check-then-create pattern.

All in all, this should reduce queries from 7-8 to 5-6 per action invocation.

On my local, I benchmarked the before/after with 100 iterations.
- New session creation: 14.5ms -> 8.3ms (-6.2ms)
- Resume session: 17.01ms -> 13.85ms (-3.16ms)
- Resume with playtime: 17.12ms -> 12.53ms (-4.59ms)

These micro-optimizations should collectively save ~25 CPU hours per month.